+++ packages/frontend/src/components/common/Dashboard/DashboardHeader.tsx
@@ -32,7 +32,7 @@ import {
     PageTitleContainer,
     SeparatorDot,
 } from '../PageHeader';
-import SpaceInfo from '../PageHeader/SpaceInfo';
+import SpaceAndDashboardInfo from '../PageHeader/SpaceAndDashboardInfo';
 import { UpdatedInfo } from '../PageHeader/UpdatedInfo';
 import ViewInfo from '../PageHeader/ViewInfo';
 import SpaceActionModal, { ActionType } from '../SpaceActionModal';
@@ -161,9 +161,11 @@ const DashboardHeader = ({
                         <>
                             <SeparatorDot icon="dot" size={6} />
 
-                            <SpaceInfo
-                                link={`/projects/${projectUuid}/spaces/${dashboardSpaceUuid}`}
-                                name={dashboardSpaceName}
+                            <SpaceAndDashboardInfo
+                                space={{
+                                    link: `/projects/${projectUuid}/spaces/${dashboardSpaceUuid}`,
+                                    name: dashboardSpaceName,
+                                }}
                             />
                         </>
                     )}